Modify: 프로필 이미지 삭제 부분 if 추가

+++ src/main/java/com/hoodie/otti/service/profile/UserProfileService.java
@@ -58,10 +58,12 @@ public class UserProfileService {
         User user = userOptional.get();
 
         String oldImageUrl = user.getProfilePhotoUrl();
-        try {
-            imageService.deleteProfileImage(oldImageUrl);
-        } catch (Exception e) {
-            throw new RuntimeException("기존 이미지 삭제에 실패했습니다: " + e.getMessage());
+        if (requestDto.getImage() != null) {
+            try {
+                imageService.deleteProfileImage(oldImageUrl);
+            } catch (Exception e) {
+                throw new RuntimeException("기존 이미지 삭제에 실패했습니다: " + e.getMessage());
+            }
         }
 
         if (requestDto.getImage() != null && !requestDto.getImage().isEmpty()) {
